fix: rework logic for models and all responses

+++ src/main/java/org/example/fitseek/repository/TrainingTypeRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TrainingTypeRepository extends JpaRepository<TrainingType, Integer> {
+public interface TrainingTypeRepository extends JpaRepository<TrainingType, Long> {
     TrainingType findByName(String name);
 }
